+++ packages/yourloops/pages/signup/signup-account-form.tsx
@@ -32,17 +32,19 @@ import { useTranslation } from "react-i18next";
 import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
 import TextField from "@material-ui/core/TextField";
+import Typography from "@material-ui/core/Typography";
+import InfoIcon from "@material-ui/icons/Info";
 
 import { errorTextFromException, REGEX_EMAIL } from "../../lib/utils";
 import { checkPasswordStrength } from "../../lib/auth/helpers";
 import metrics from "../../lib/metrics";
-import SignUpFormProps from "./signup-form-props";
 import { useAuth } from "../../lib/auth";
 import { getCurrentLang } from "../../lib/language";
 import { useAlert } from "../../components/utils/snackbar";
-import { useSignUpFormState } from "./signup-formstate-context";
 import { PasswordStrengthMeter } from "../../components/utils/password-strength-meter";
 import Password from "../../components/utils/password";
+import { useSignUpFormState } from "./signup-formstate-context";
+import SignUpFormProps from "./signup-form-props";
 
 
 interface Errors {
@@ -149,6 +151,17 @@ function SignUpAccountForm(props: SignUpFormProps): JSX.Element {
         margin="normal"
         required
       />
+      <Box display="flex" m={0}>
+        <InfoIcon color="primary" />
+        <Typography
+          id="signup-account-deletion-warning"
+          variant="body2"
+          style={{ textAlign: "left", marginLeft: "10px" }}
+        >
+          {t("signup-account-deletion-warning")}
+        </Typography>
+      </Box>
+
       <Box
         id="signup-accountform-button-group"
         display="flex"