+++ packages/yourloops/test/pages/stepper/index.ts
@@ -27,9 +27,11 @@
  */
 
 import TestSignupProfileForm from "./signup-profile-form.test";
+import TestSignupAccountForm from "./signup-account-form.test";
 
 function testSignupProfilePage(): void {
   describe("Profile form", TestSignupProfileForm);
+  describe("Account form", TestSignupAccountForm);
 }
 
 export default testSignupProfilePage;