+++ CHANGELOG.md
@@ -7,6 +7,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-1105 Yourloops new blue theme
 - YLP-1122 Patient should enter the team unique ID to accept a team invitation
 - YLP-1188 Pink theme for clinical environment
+- YLP-1190 Advertise user about account deletion
 ### Improvements
 - YLP-22 Print PDF content cleaned and improved (removing useless data)
 - YLP-1120 Overview tab cleanup